Multiview support (#1408)
* fixing layered textures for vulkan

* fix offsets for compressed textures

* fix mipmap generation for cubemaps

* creating subclasses from glrender-texture

* adding get-set methods

* rearchitect renderer

* fixing bugs for posteffects

* Fix: don't update mFormat while updating from Buffer (#1361)

Fix for: update texture from NIO buffer #1316, #1173
Problem: GLBitmapImeage::updateFromBuffer() return type is int,
but method body doesn't have return statement
Solution: Don't update mFormat.
For buffers, format is a parameter to the update() method.

* framework: propagate all exceptions openStream might throw (#1353)

* [DockListener] separated addDockListener from the constructor (#1354)

- changed it so that constructor won't call it. Instead it's called separately.


* vulkan: framework, backend_oculus: add support for building with clang (#1359)

* framework, backend_oculus: add support for building with clang

* framework: minor code changes to fix some of the crashes when compiled with clang

+ makefile updates specific to vulkan

* Calculate UV, normal, and Barycentric coords for picked Mesh Colliders (#1360)

* vulkan: fixes gvr-renderableview (#1366)

* particle system for vulkan branch + enable/disable depth writing (#1340)

* particle system for vulkan branch + enable/disable depth writing from application

* cleanup

* cleanup + changes to work with vulkan

* enable/disable depth writing in vulkan

* correctly clear up array holding particle meshes

* Remove EyePointee references (#1362)

* vulkan: remove loadFutureMesh call from 3dcursor extension (#1368)

* fix post-effect rendering

* fix oes shaders for vulkan

* merge Nola's changes

+++ GVRf/Framework/framework/src/main/jni/objects/textures/image.h
@@ -160,8 +160,8 @@ protected:
     std::mutex mUpdateLock;
     short   mType;
     short   mLevels;
-    short   mWidth;
-    short   mHeight;
+    int   mWidth;
+    int   mHeight;
     short   mDepth;
     short   mState;
     int     mImageSize;
